+++ src/ralph/ui/views/racks.py
@@ -16,7 +16,7 @@ from django.contrib import messages
 
 from ralph.discovery.models import ReadOnlyDevice, Device, DeviceType
 from ralph.ui.views.common import (Info, Prices, Addresses, Costs,
-    Purchase, Components, History, Discover)
+    Purchase, Components, History, Discover, Reports)
 from ralph.account.models import Perm
 from ralph.util import presentation
 from ralph.ui.views.common import BaseMixin
@@ -87,35 +87,50 @@ class SidebarRacks(object):
         })
         return ret
 
+
 class Racks(SidebarRacks, BaseMixin):
     pass
+
 
 class RacksInfo(Racks, Info):
     pass
 
+
 class RacksAddresses(Racks, Addresses):
     pass
+
 
 class RacksComponents(Racks, Components):
     pass
 
+
 class RacksCMDB(Racks, CMDB,DeviceDetailView ):
     pass
+
 
 class RacksPrices(Racks, Prices):
     pass
 
+
 class RacksCosts(Racks, Costs):
     pass
+
 
 class RacksHistory(Racks, History):
     pass
 
+
 class RacksPurchase(Racks, Purchase):
     pass
 
+
 class RacksDiscover(Racks, Discover):
     pass
+
+
+class RacksReports(Racks, Reports):
+    pass
+
 
 class RacksDeviceList(SidebarRacks, BaseMixin, BaseDeviceList):
     def user_allowed(self):
@@ -185,6 +200,7 @@ class RacksDeviceList(SidebarRacks, BaseMixin, BaseDeviceList):
             'subsection_slug': self.rack.sn if self.rack else self.rack,
         })
         return ret
+
 
 class DeviceCreateView(CreateView):
     model = Device